don't use HEAD connectivity test to determine a package URL is reachable, fix lots of expected Join-Path errors flooding ERROR variable
The URL connectivity test with a HEAD request is not a perfectly
realistic test and just adds overhead - I can just fail later when I try
to download/GET the file and it doesn't work.

The only files I use the reachability test for now are the model or
database XML packages index files, so I increased the timeout for the
HEAD request as well.