fix Install-LSUpdate redownloading installers from http repositories when they were already previously downloaded
This new regression was particularly problematic for environments that
require proxy settings for LSUClient to access the HTTP repository. In
previous versions of LSUClient, Install-LSUpdate did not have any proxy
parameters and existing scripts will not expect it to try to connect to
the internet when all needed filess were previously downloaded with e.g.
Save-LSUpdate (which did take proxy parameters)