use up to 5 characters of the computer model number for repository lookup - fixes #38
after experimenting with different model/MTM IDs, this seems to be the
behavior of the official system update: reject lengths below 4, use 4
and 5 as-is and cut to the first 4 characters if it's longer than that.

I have also (re)moved the parameter-validation on 'Model' because it is
now validated later in code. This allows one to pass full-length model
codes to Get-LSUpdate and have it auto-trim them appropriately.