replace OpenWindows property, Write-Host with output to appropriate streams
the information about any open windows is for troubleshooting purposes,
making it accessible programmatically encourages scripting based on it
which I don't think is ideal. It also wouldn't format in a readable way
unless pre-accounted for with custom output of the nested properties.

also general go-live fixups of moving output to the most appropriate
streams instead of the temp Write-Host calls