revert 51a79de to not risk launching a wrong executable
unconditionally removing pathParts from the beginning of a
command could lead to a wrong file being executed, especially
if the intended file cannot be found (e.g. failed download)
and now that we're also searching for executables on PATH.

Attempting to parse cmd START commands is really complex and
would result in the same, if not more problems. Revisit later.