add support for _Coreq dependencies to fix #58
Different approach than the initial attempt 6b26c1c.
Package dependencies are now processed later, after all other aspects of
all packages have already been processed. This allows us to look up the
IsInstalled property of all packages during dependency tests which is
needed for Coreq.

This works because DetectInstall tests cannot contain Coreq tests.

Support is not perfect, a Coreq dependency may be on a lower version
of a package than what is currently available in the repository. If
the older version is installed but the latest isn't the Coreq test
will currently fail. Will need to evaluate DetectVersion to fix this.