fail out of AND tests as soon as 1 came back false
This could fix / work around issue #103 because the problematic package:
https://download.lenovo.com/pccbbs/mobiles/n3acv41w_2_.xml
has lots of idependency tests in a long AND-clause and the one that's
hanging is the very last one. So if any previous test had already failed
we could just skip all the others, thus avoiding the problematic
ExternalDetection with MCUFWRevCheck.exe.